requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
mysql-connector-repackaged==0.3.1
Werkzeug==3.1.3
gunicorn==23.0.0